+++ eslint-config/src/cli.js
@@ -3,9 +3,9 @@
 
 /** @type {import("eslint").Linter.Config} */
 const config = {
-	extends: "plugin:@redguy12/node",
+	extends: require.resolve("./node.js"),
 	parserOptions: { ecmaVersion: "latest" },
 	rules: { "fp/no-throw": 0, "no-console": 0 },
 };
 
-module.exports = config;
+module.exports = (config);